Broke Thesaurus Definition Popover to its own component
