+++ src/pages/SearchPage/SearchPage.jsx
@@ -9,21 +9,14 @@ import {
   Box,
   chakra,
   Fade,
-  Popover,
-  PopoverTrigger,
-  PopoverContent,
-  PopoverHeader,
-  PopoverBody,
-  PopoverArrow,
-  PopoverCloseButton,
   useColorModeValue,
-  IconButton,
   useMediaQuery,
   Button,
 } from '@chakra-ui/react';
-import { InfoIcon, SearchIcon } from '@chakra-ui/icons';
+import { SearchIcon } from '@chakra-ui/icons';
 import { connect } from 'react-redux';
 import SearchInput from '../../components/search-input/search-input.component';
+import ThesaurusDefinitionPopover from '../../components/thesaurus-definition/thesaurus-definition.component';
 
 const MainContainer = tw.div`
    text-center
@@ -36,9 +29,6 @@ const SearchPage = ({
   WordFindType, WordFind, isOpen, HandleSearchButtonClick, getWords,
 }) => {
   const [isMoreThan420px] = useMediaQuery('(max-width: 420px)');
-  const fontColorDarkWhiteSmallWords = useColorModeValue('#3B82F6', 'orange');
-  const hover = useColorModeValue({ color: '#3B82F6' }, { color: 'orange' });
-  const fontColorMain = useColorModeValue('#edf2f7', 'gray.800');
   const focusBorderColorGeneral = useColorModeValue({ boxShadow: '0 0 0 3px #3B82F6' }, { boxShadow: '0 0 0 3px orange' });
 
   return (
@@ -48,22 +38,7 @@ const SearchPage = ({
           Words to be
           {(isMoreThan420px) ? '\u00A0' : ' thesaurused'}
         </Heading>
-        <Popover placement="auto-end" display="inline-block">
-          <PopoverTrigger>
-            {(isMoreThan420px)
-              ? <Heading textDecoration={`underline wavy ${fontColorDarkWhiteSmallWords}`} _hover={[null, null, hover]} cursor="pointer" fontFamily="Playfair Display" fontSize={['3xl', '5xl', '6xl']} whiteSpace="nowrap">thesaurused</Heading>
-              : <IconButton variant="ghost" outline="none" outlineColor="initial" style={{ outlineStyle: 'none' }} _focus={focusBorderColorGeneral} ml="2" size="sm" icon={<InfoIcon w="5" h="5" />} />}
-          </PopoverTrigger>
-          <PopoverContent outline="none" outlineColor="initial" style={{ outlineStyle: 'none' }} _focus={focusBorderColorGeneral} background={fontColorMain}>
-            <PopoverArrow />
-            <PopoverCloseButton />
-            <PopoverHeader fontFamily="Playfair Display" fontSize={['xl']} fontWeight="bold">Thesaurus</PopoverHeader>
-            <PopoverBody>
-              A thesaurus (plural thesauri or thesauruses) or synonym dictionary
-              is a reference work for finding synonyms and sometimes antonyms of words.
-            </PopoverBody>
-          </PopoverContent>
-        </Popover>
+        <ThesaurusDefinitionPopover />
       </Box>
       <Box display={{ sm: 'flex', md: 'flex' }} justifyContent={[null, 'center', null]}>
         <Box>
